backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Production Server
gunicorn>=21.2.0

# Environment & Configuration
python-dotenv>=1.0.0

# LangChain - AI Agent Framework
langchain==0.1.20
langchain-core>=0.1.52
langchain-groq>=0.1.5

# LLM Provider
groq>=0.4.0

# Web Search APIs
tavily-python>=0.3.0
google-search-results>=2.4.2

# HTTP Client
requests>=2.31.0

# Database
sqlalchemy>=2.0.0

# Math & Calculations
sympy>=1.12

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Timezone Support
pytz>=2024.1
